[HOT-FIX] fixed issue compiling 'uint' type definitions on Windows
Former-commit-id: 2030206bd9e22364949984911083ce5cdc6c8656
Former-commit-id: 481550c8db4b2ef8b08c4d1ba6a1ee90662402ed

+++ src/simulation/dynamics/facetDragEffector/facetDragDynamicEffector.cpp
@@ -150,7 +150,7 @@ void FacetDragDynamicEffector::plateDrag(){
     this->forceExternal_B.setZero();
     this->torqueExternalPntB_B.setZero();
     
-	for(uint i = 0; i < this->numFacets; i++){
+	for(size_t i = 0; i < this->numFacets; i++){
 	    projectionTerm = this->scGeometry.facetNormals_B[i].dot(this->v_hat_B);
 		projectedArea = this->scGeometry.facetAreas[i] * projectionTerm;
 		if(projectedArea > 0.0){
